+++ IceEmul/iceEmul.py
@@ -21,37 +21,65 @@ def scatterplot_ice(ax, var, varname, bounds=None, cmap='viridis'):
     ax.set_title(varname)
     plt.colorbar(scatter, ax=ax, orientation='horizontal', shrink=0.5, pad=0.02)
 
-# Load the NetCDF file
-file_path = "gdas.t00z.icef009.ffnn.nc.20210710"
-ds = xr.open_dataset(file_path)
+class Ice:
+    def __init__(self, filename):
+        self.filename = filename
+        ds = xr.open_dataset(filename)
+        self.aice = ds["aice"].values
+        self.aice_ffnn = ds["aice_ffnn"].values
+        self.dcdt = ds["dcdt"].values
+        self.dcds = ds["dcds"].values
+        self.dcdsi = ds["dcdsi"].values
+        self.dcdhi = ds["dcdhi"].values
+        self.dcdhs = ds["dcdhs"].values
+        self.dcdtair = ds["dcdtair"].values
+        self.dcdtsfc = ds["dcdtsfc"].values
+        self.lon = ds["lon"].values
+        self.lat = ds["lat"].values
 
-# Extract the variables
-aice = ds["aice"].values
-aice_ffnn = ds["aice_ffnn"].values
-dcdt = ds["dcdt"].values
-dcds = ds["dcds"].values
-dcdsi = ds["dcdsi"].values
-dcdhi = ds["dcdhi"].values
-dcdhs = ds["dcdhs"].values
-dcdtair = ds["dcdtair"].values
-dcdtsfc = ds["dcdtsfc"].values
-lon = ds["lon"].values
-lat = ds["lat"].values
+    def diff(self, other):
+        self.aice -= other.aice
+        self.aice_ffnn -= other.aice_ffnn
+
+    def scatterplot_ice(self, ax, var, varname, bounds=None, cmap='viridis'):
+        if (bounds is None):
+            scatter = ax.scatter(self.lon, self.lat, c=var,
+                                 cmap=cmap, s=.5, alpha=0.8, transform=ccrs.PlateCarree())
+        else:
+            scatter = ax.scatter(self.lon, self.lat, c=var,
+                                 cmap=cmap, s=.5, alpha=0.8, transform=ccrs.PlateCarree(),
+                                 vmin=bounds[0], vmax=bounds[1])
+        ax.coastlines()
+        ax.set_title(varname)
+        plt.colorbar(scatter, ax=ax, orientation='horizontal', shrink=0.5, pad=0.02)
+
+
+
+# Load the NetCDF file
+f009 = Ice("gdas.t00z.icef009.ffnn.nc")
+f003 = Ice("gdas.t00z.icef003.ffnn.nc")
+f009.diff(f003)
 
 # plot stuff
 fig = plt.figure(figsize=(8, 8))
-gs = GridSpec(3, 3, width_ratios=[1, 1, 1], height_ratios=[1, 1, 1])
+#gs = GridSpec(3, 3, width_ratios=[1, 1, 1], height_ratios=[1, 1, 1])
+gs = GridSpec(1, 2, width_ratios=[1, 1], height_ratios=[1])
 
 #projection = ccrs.SouthPolarStereo()
 projection = ccrs.NorthPolarStereo()
 
 # Target aice
 ax1 = fig.add_subplot(gs[0, 0], projection=projection)
-scatterplot_ice(ax1, aice, "$aice$", bounds=[0, 1])
+f009.scatterplot_ice(ax1, f009.aice, "$aice$", bounds=[-0.1, 0.1], cmap='RdBu_r')
+#f009.scatterplot_ice(ax1, f009.aice, "$aice$", bounds=[0, 1])
 
 # Target aice
 ax2 = fig.add_subplot(gs[0, 1], projection=projection)
-scatterplot_ice(ax2, aice_ffnn, "$aice_{ffnn}$", bounds=[0, 1])
+f009.scatterplot_ice(ax2, f009.aice_ffnn, "$aice_{ffnn}$", bounds=[-0.1, 0.1], cmap='RdBu_r')
+#f009.scatterplot_ice(ax2, f009.aice_ffnn, "$aice_{ffnn}$", bounds=[0, 1])
+
+plt.tight_layout()
+plt.show()
 
 # Un-balanced aice
 #ax2 = fig.add_subplot(gs[0, 2], projection=projection)
@@ -59,31 +87,31 @@ scatterplot_ice(ax2, aice_ffnn, "$aice_{ffnn}$", bounds=[0, 1])
 
 # dcdsst
 ax3 = fig.add_subplot(gs[1, 0], projection=projection)
-scatterplot_ice(ax3, dcdt, "$dc/dT_{ocean}$", cmap='RdBu', bounds=[-0.3, 0.3])
+f009.scatterplot_ice(ax3, f009.dcdt, "$dc/dT_{ocean}$", cmap='RdBu', bounds=[-0.3, 0.3])
 
 # dcds
 ax4 = fig.add_subplot(gs[1, 1], projection=projection)
-scatterplot_ice(ax4, dcds, "$dc/dS_{ocean}$", cmap='RdBu', bounds=[-0.02, 0.02])
+f009.scatterplot_ice(ax4, f009.dcds, "$dc/dS_{ocean}$", cmap='RdBu', bounds=[-0.02, 0.02])
 
 # dcdsi
 ax = fig.add_subplot(gs[0, 2], projection=projection)
-scatterplot_ice(ax, dcdsi, "$dc/dS_{ice}$", cmap='RdBu', bounds=[-0.05, 0.05])
+f009.scatterplot_ice(ax, f009.dcdsi, "$dc/dS_{ice}$", cmap='RdBu', bounds=[-0.05, 0.05])
 
 # dcdhi
 ax5 = fig.add_subplot(gs[1, 2], projection=projection)
-scatterplot_ice(ax5, dcdhi, "$dc/dh_{ice}$", cmap='RdBu', bounds=[-3, 3])
+f009.scatterplot_ice(ax5, f009.dcdhi, "$dc/dh_{ice}$", cmap='RdBu', bounds=[-3, 3])
 
 # dcdhs
 ax6 = fig.add_subplot(gs[2, 0], projection=projection)
-scatterplot_ice(ax6, dcdhs, "$dc/dh_{snow}$", cmap='RdBu', bounds=[-3, 3])
+f009.scatterplot_ice(ax6, f009.dcdhs, "$dc/dh_{snow}$", cmap='RdBu', bounds=[-3, 3])
 
 # dcdtsfc
 ax = fig.add_subplot(gs[2, 1], projection=projection)
-scatterplot_ice(ax, dcdtsfc, "$dc/dT_{sfc}$", cmap='RdBu', bounds=[-0.1, 0.1])
+f009.scatterplot_ice(ax, f009.dcdtsfc, "$dc/dT_{sfc}$", cmap='RdBu', bounds=[-0.1, 0.1])
 
 # dcdtair
 ax = fig.add_subplot(gs[2, 2], projection=projection)
-scatterplot_ice(ax, dcdtair, "$dc/dT_{air}$", cmap='RdBu', bounds=[-0.05, 0.05])
+f009.scatterplot_ice(ax, f009.dcdtair, "$dc/dT_{air}$", cmap='RdBu', bounds=[-0.05, 0.05])
 
 plt.tight_layout()
 plt.show()
