+++ test/daml_ice_emul.yaml
@@ -5,9 +5,9 @@ ffnn:
   load model: ./arctic/model.pt
 
 prediction:
-  # cice history: "/home/gvernier/data/20210715/00/gdas.t00z.icef009.nc"
-  cice history: "/home/gvernier/data/20210710/00/gdas.t00z.icef003.nc"
-  output filename: "gdas.t00z.icef009.ffnn.nc.20210710"
+  cice history: "/home/gvernier/data/20210715/00/gdas.t00z.icef003.nc"
+  #cice history: "/home/gvernier/data/20210710/00/gdas.t00z.icef003.nc"
+  output filename: "gdas.t00z.icef003.ffnn.nc"
 
 #training:
 #  epochs: 200
